requirements.txt
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
celery==5.2.7
django-environ==0.9.0
